+++ CHANGELOG.md
@@ -46,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Updated runtime to 0.6.0-rc2
 - Updated hdremix to a1863ffe
+- Renamed `Feature Flags` to `Experimental Features`
 
 ### Fixed
 - REMIX-3401: Fixed hot-reload by allowing reuse of validators