+++ source/extensions/lightspeed.trex.menu.workfile/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.1]
+### Changed
+- Rename `Feature Flags` to `Experimental Features`
+
 ## [1.1.0]
 ### Added
 - Added `Feature Flags` entry to the menu